+++ blade-graphics/src/vulkan/resource.rs
@@ -459,7 +459,7 @@ impl crate::traits::ResourceDevice for super::Context {
         }
     }
 
-    fn sync_buffer(&self, _buffer: super::Buffer, _target: crate::BufferTarget) {}
+    fn sync_buffer(&self, _piece: crate::BufferPiece, _size: u64, _target: crate::BufferTarget) {}
 
     fn destroy_buffer(&self, buffer: super::Buffer) {
         log::info!(