+++ examples/bunnymark/example.rs
@@ -132,7 +132,11 @@ impl Example {
                 texture_data.len(),
             );
         }
-        context.sync_buffer(upload_buffer, gpu::BufferTarget::Data);
+        context.sync_buffer(
+            upload_buffer.into(),
+            upload_buffer.size(),
+            gpu::BufferTarget::Data,
+        );
 
         let sampler = context.create_sampler(gpu::SamplerDesc {
             name: "main",
@@ -157,7 +161,11 @@ impl Example {
                 vertex_data.len(),
             );
         }
-        context.sync_buffer(vertex_buf, gpu::BufferTarget::Data);
+        context.sync_buffer(
+            vertex_buf.into(),
+            vertex_buf.size(),
+            gpu::BufferTarget::Data,
+        );
 
         let bunnies = vec![Sprite {
             data: SpriteData {